修改接口

+++ mktcenter-serviceimpl/src/main/java/com/bizvane/mktcenterserviceimpl/service/impl/TaskServiceImpl.java
@@ -372,8 +372,8 @@ public class TaskServiceImpl implements TaskService {
                 }
                 ResponseData<CouponFindCouponCountResponseVO> couponCount= couponQueryService.findCouponCountBySendBusinessId(task.getTaskId(), sendType, sysBrandId);
                 CouponFindCouponCountResponseVO data = couponCount.getData();
-                int couponSum = data.getCouponSum();
-                task.setOneTaskInvalidCountCoupon(Long.valueOf(couponSum));
+                Long couponSum = data.getCouponSum();
+                task.setOneTaskInvalidCountCoupon(couponSum);
 
                 allPoints=allPoints+task.getOneTaskPoints();
                 allCountCoupon= allCountCoupon+task.getOneTaskCountCoupon();
